_data/requirements.txt
alembic==1.0.10
amqp==2.5.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
colorlog==4.0.2
filetype==1.0.5
Flask==1.0.3
Flask-Login==0.4.1
Flask-Migrate==2.5.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
har2case==0.2.0
HttpRunner==2.1.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.6.1
Mako==1.0.10
MarkupSafe==1.1.1
PyMySQL==0.9.3
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
requests-toolbelt==0.9.1
selenium==3.141.0
six==1.12.0
SQLAlchemy==1.3.3
urllib3==1.25.3
vine==1.3.0
Werkzeug==0.15.4
WTForms==2.2.1
